Move config class
