+++ dask_cloudprovider/config.py
@@ -6,6 +6,19 @@ import dask
 import yaml
 
 
+class ClusterConfig(dict):
+    """Simple config interface for dask-cloudprovider clusters, such as `AzureVMCluster`.
+
+    Enables '.' notation for nested access, as per `dask.config.get`.
+
+    """
+    def __new__(cls, d):
+        return super().__new__(cls, d)
+        
+    def get(self, key, default=None, override_with=None):
+        return dask.config.get(key, default=default, config=self, override_with=override_with)
+
+
 fn = os.path.join(os.path.dirname(__file__), "cloudprovider.yaml")
 dask.config.ensure_file(source=fn)
